Adds support for auto-and predicates
Augments hexagon.slaspec to split up newcmpjump instructions into the
compare (part1) and dot-new jumps (part2) based on the content of two
context registers

Reworks the logic in HexagonPcodeEmitPacked to split up newcmpjumps and
updates a comment describing what's happening

Adds a test to HexagonPacketTestDecompilation

Moves some constant maps and sets around

Fixes #4
