+++ Ghidra/Processors/Hexagon/src/main/java/ghidra/app/plugin/core/analysis/HexagonPacketAnalyzer.java
@@ -276,15 +276,6 @@ public class HexagonPacketAnalyzer extends AbstractAnalyzer {
 				}
 			}
 
-			try {
-				// perform some validation of the packet for constructs we don't support
-				insnIter = program.getListing().getInstructions(addrSet, true);
-				packet.validatePredicates(program, insnIter);
-			} catch (UnknownInstructionException e) {
-				Msg.error(this, "Packet at " + packet.packetStartAddress + " had unimplemented construct " + e);
-				throw e;
-			}
-
 			AutoAnalysisManager.getAnalysisManager(program).codeDefined(disassembled);
 
 		} catch (UnknownInstructionException e) {
