+++ Ghidra/Processors/Hexagon/src/main/java/ghidra/app/plugin/core/analysis/HexagonPcodeEmitPacked.java
@@ -18,19 +18,29 @@ package ghidra.app.plugin.core.analysis;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import ghidra.app.plugin.processors.sleigh.PcodeEmitPacked;
+import ghidra.app.util.PseudoInstruction;
 import ghidra.program.model.address.Address;
+import ghidra.program.model.address.AddressOverflowException;
 import ghidra.program.model.address.AddressSet;
 import ghidra.program.model.address.AddressSpace;
 import ghidra.program.model.address.UniqueAddressFactory;
 import ghidra.program.model.lang.InstructionContext;
+import ghidra.program.model.lang.InstructionPrototype;
+import ghidra.program.model.lang.InsufficientBytesException;
 import ghidra.program.model.lang.PackedBytes;
+import ghidra.program.model.lang.ProcessorContext;
+import ghidra.program.model.lang.ProcessorContextImpl;
 import ghidra.program.model.lang.Register;
 import ghidra.program.model.lang.UnknownInstructionException;
+import ghidra.program.model.listing.ContextChangeException;
 import ghidra.program.model.listing.FlowOverride;
 import ghidra.program.model.listing.Instruction;
 import ghidra.program.model.listing.InstructionIterator;
@@ -43,6 +53,9 @@ public class HexagonPcodeEmitPacked {
 
 	Program program;
 
+	Register part1Register;
+	Register part2Register;
+
 	Address minAddr;
 	Address maxAddr;
 	Address pktNext;
@@ -54,6 +67,9 @@ public class HexagonPcodeEmitPacked {
 
 	HexagonPcodeEmitPacked(Program program) {
 		this.program = program;
+		part1Register = program.getProgramContext().getRegister("part1");
+		part2Register = program.getProgramContext().getRegister("part2");
+
 	}
 
 	boolean isCallotherNewreg(PcodeOp op) {
@@ -233,6 +249,254 @@ public class HexagonPcodeEmitPacked {
 		}
 	}
 
+	static final Map<String, Integer> dot_new_predicates;
+
+	static {
+		dot_new_predicates = new HashMap<>();
+		dot_new_predicates.put("J4_cmpeqi_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpeqi_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpeqi_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpeqi_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgti_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgti_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgti_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgti_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgtui_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgtui_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgtui_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgtui_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpeqn1_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpeqn1_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpeqn1_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpeqn1_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgtn1_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgtn1_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgtn1_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgtn1_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_tstbit0_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_tstbit0_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_tstbit0_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_tstbit0_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpeq_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpeq_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpeq_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpeq_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgt_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgt_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgt_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgt_fp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgtu_tp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgtu_fp0_jump_nt", 0);
+		dot_new_predicates.put("J4_cmpgtu_tp0_jump_t", 0);
+		dot_new_predicates.put("J4_cmpgtu_fp0_jump_t", 0);
+		dot_new_predicates.put("SA1_clrtnew", 0);
+		dot_new_predicates.put("SA1_clrfnew", 0);
+		dot_new_predicates.put("SL2_return_tnew", 0);
+		dot_new_predicates.put("SL2_return_fnew", 0);
+		dot_new_predicates.put("SL2_jumpr31_tnew", 0);
+		dot_new_predicates.put("SL2_jumpr31_fnew", 0);
+		dot_new_predicates.put("J4_cmpeqi_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpeqi_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpeqi_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpeqi_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgti_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgti_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgti_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgti_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgtui_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgtui_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgtui_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgtui_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpeqn1_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpeqn1_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpeqn1_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpeqn1_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgtn1_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgtn1_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgtn1_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgtn1_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_tstbit0_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_tstbit0_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_tstbit0_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_tstbit0_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpeq_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpeq_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpeq_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpeq_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgt_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgt_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgt_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgt_fp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgtu_tp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgtu_fp1_jump_nt", 1);
+		dot_new_predicates.put("J4_cmpgtu_tp1_jump_t", 1);
+		dot_new_predicates.put("J4_cmpgtu_fp1_jump_t", 1);
+	}
+
+	static final Set<String> dot_new_predicates_operands;
+
+	static {
+		dot_new_predicates_operands = new HashSet<>();
+		dot_new_predicates_operands.add("J2_jumptnew");
+		dot_new_predicates_operands.add("J2_jumpfnew");
+		dot_new_predicates_operands.add("J2_jumptnewpt");
+		dot_new_predicates_operands.add("J2_jumpfnewpt");
+		dot_new_predicates_operands.add("J2_jumprtnew");
+		dot_new_predicates_operands.add("J2_jumprfnew");
+		dot_new_predicates_operands.add("J2_jumprtnewpt");
+		dot_new_predicates_operands.add("J2_jumprfnewpt");
+		dot_new_predicates_operands.add("L4_return_tnew_pt");
+		dot_new_predicates_operands.add("L4_return_fnew_pt");
+		dot_new_predicates_operands.add("L4_return_tnew_pnt");
+		dot_new_predicates_operands.add("L4_return_fnew_pnt");
+		dot_new_predicates_operands.add("L2_ploadrubtnew_io");
+		dot_new_predicates_operands.add("L2_ploadrubfnew_io");
+		dot_new_predicates_operands.add("L4_ploadrubtnew_rr");
+		dot_new_predicates_operands.add("L4_ploadrubfnew_rr");
+		dot_new_predicates_operands.add("L2_ploadrubtnew_pi");
+		dot_new_predicates_operands.add("L2_ploadrubfnew_pi");
+		dot_new_predicates_operands.add("L4_ploadrubtnew_abs");
+		dot_new_predicates_operands.add("L4_ploadrubfnew_abs");
+		dot_new_predicates_operands.add("L2_ploadrbtnew_io");
+		dot_new_predicates_operands.add("L2_ploadrbfnew_io");
+		dot_new_predicates_operands.add("L4_ploadrbtnew_rr");
+		dot_new_predicates_operands.add("L4_ploadrbfnew_rr");
+		dot_new_predicates_operands.add("L2_ploadrbtnew_pi");
+		dot_new_predicates_operands.add("L2_ploadrbfnew_pi");
+		dot_new_predicates_operands.add("L4_ploadrbtnew_abs");
+		dot_new_predicates_operands.add("L4_ploadrbfnew_abs");
+		dot_new_predicates_operands.add("L2_ploadruhtnew_io");
+		dot_new_predicates_operands.add("L2_ploadruhfnew_io");
+		dot_new_predicates_operands.add("L4_ploadruhtnew_rr");
+		dot_new_predicates_operands.add("L4_ploadruhfnew_rr");
+		dot_new_predicates_operands.add("L2_ploadruhtnew_pi");
+		dot_new_predicates_operands.add("L2_ploadruhfnew_pi");
+		dot_new_predicates_operands.add("L4_ploadruhtnew_abs");
+		dot_new_predicates_operands.add("L4_ploadruhfnew_abs");
+		dot_new_predicates_operands.add("L2_ploadrhtnew_io");
+		dot_new_predicates_operands.add("L2_ploadrhfnew_io");
+		dot_new_predicates_operands.add("L4_ploadrhtnew_rr");
+		dot_new_predicates_operands.add("L4_ploadrhfnew_rr");
+		dot_new_predicates_operands.add("L2_ploadrhtnew_pi");
+		dot_new_predicates_operands.add("L2_ploadrhfnew_pi");
+		dot_new_predicates_operands.add("L4_ploadrhtnew_abs");
+		dot_new_predicates_operands.add("L4_ploadrhfnew_abs");
+		dot_new_predicates_operands.add("L2_ploadritnew_io");
+		dot_new_predicates_operands.add("L2_ploadrifnew_io");
+		dot_new_predicates_operands.add("L4_ploadritnew_rr");
+		dot_new_predicates_operands.add("L4_ploadrifnew_rr");
+		dot_new_predicates_operands.add("L2_ploadritnew_pi");
+		dot_new_predicates_operands.add("L2_ploadrifnew_pi");
+		dot_new_predicates_operands.add("L4_ploadritnew_abs");
+		dot_new_predicates_operands.add("L4_ploadrifnew_abs");
+		dot_new_predicates_operands.add("L2_ploadrdtnew_io");
+		dot_new_predicates_operands.add("L2_ploadrdfnew_io");
+		dot_new_predicates_operands.add("L4_ploadrdtnew_rr");
+		dot_new_predicates_operands.add("L4_ploadrdfnew_rr");
+		dot_new_predicates_operands.add("L2_ploadrdtnew_pi");
+		dot_new_predicates_operands.add("L2_ploadrdfnew_pi");
+		dot_new_predicates_operands.add("L4_ploadrdtnew_abs");
+		dot_new_predicates_operands.add("L4_ploadrdfnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerbtnew_io");
+		dot_new_predicates_operands.add("S4_pstorerbfnew_io");
+		dot_new_predicates_operands.add("S4_pstorerbtnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerbfnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerbtnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerbfnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerbtnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerbfnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerhtnew_io");
+		dot_new_predicates_operands.add("S4_pstorerhfnew_io");
+		dot_new_predicates_operands.add("S4_pstorerhtnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerhfnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerhtnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerhfnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerhtnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerhfnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerftnew_io");
+		dot_new_predicates_operands.add("S4_pstorerffnew_io");
+		dot_new_predicates_operands.add("S4_pstorerftnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerffnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerftnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerffnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerftnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerffnew_abs");
+		dot_new_predicates_operands.add("S4_pstoreritnew_io");
+		dot_new_predicates_operands.add("S4_pstorerifnew_io");
+		dot_new_predicates_operands.add("S4_pstoreritnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerifnew_rr");
+		dot_new_predicates_operands.add("S2_pstoreritnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerifnew_pi");
+		dot_new_predicates_operands.add("S4_pstoreritnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerifnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerdtnew_io");
+		dot_new_predicates_operands.add("S4_pstorerdfnew_io");
+		dot_new_predicates_operands.add("S4_pstorerdtnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerdfnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerdtnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerdfnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerdtnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerdfnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerinewtnew_io");
+		dot_new_predicates_operands.add("S4_pstorerinewfnew_io");
+		dot_new_predicates_operands.add("S4_pstorerinewtnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerinewfnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerinewtnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerinewfnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerinewtnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerinewfnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerbnewtnew_io");
+		dot_new_predicates_operands.add("S4_pstorerbnewfnew_io");
+		dot_new_predicates_operands.add("S4_pstorerbnewtnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerbnewfnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerbnewtnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerbnewfnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerbnewtnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerbnewfnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerhnewtnew_io");
+		dot_new_predicates_operands.add("S4_pstorerhnewfnew_io");
+		dot_new_predicates_operands.add("S4_pstorerhnewtnew_rr");
+		dot_new_predicates_operands.add("S4_pstorerhnewfnew_rr");
+		dot_new_predicates_operands.add("S2_pstorerhnewtnew_pi");
+		dot_new_predicates_operands.add("S2_pstorerhnewfnew_pi");
+		dot_new_predicates_operands.add("S4_pstorerhnewtnew_abs");
+		dot_new_predicates_operands.add("S4_pstorerhnewfnew_abs");
+		dot_new_predicates_operands.add("S4_storeirbtnew_io");
+		dot_new_predicates_operands.add("S4_storeirbfnew_io");
+		dot_new_predicates_operands.add("S4_storeirhtnew_io");
+		dot_new_predicates_operands.add("S4_storeirhfnew_io");
+		dot_new_predicates_operands.add("S4_storeiritnew_io");
+		dot_new_predicates_operands.add("S4_storeirifnew_io");
+		dot_new_predicates_operands.add("C2_cmovenewit");
+		dot_new_predicates_operands.add("C2_cmovenewif");
+		dot_new_predicates_operands.add("C2_ccombinewnewt");
+		dot_new_predicates_operands.add("C2_ccombinewnewf");
+		dot_new_predicates_operands.add("A2_paddtnew");
+		dot_new_predicates_operands.add("A2_paddfnew");
+		dot_new_predicates_operands.add("A2_psubtnew");
+		dot_new_predicates_operands.add("A2_psubfnew");
+		dot_new_predicates_operands.add("A2_padditnew");
+		dot_new_predicates_operands.add("A2_paddifnew");
+		dot_new_predicates_operands.add("A2_pxortnew");
+		dot_new_predicates_operands.add("A2_pxorfnew");
+		dot_new_predicates_operands.add("A2_pandtnew");
+		dot_new_predicates_operands.add("A2_pandfnew");
+		dot_new_predicates_operands.add("A2_portnew");
+		dot_new_predicates_operands.add("A2_porfnew");
+		dot_new_predicates_operands.add("A4_psxtbtnew");
+		dot_new_predicates_operands.add("A4_psxtbfnew");
+		dot_new_predicates_operands.add("A4_pzxtbtnew");
+		dot_new_predicates_operands.add("A4_pzxtbfnew");
+		dot_new_predicates_operands.add("A4_psxthtnew");
+		dot_new_predicates_operands.add("A4_psxthfnew");
+		dot_new_predicates_operands.add("A4_pzxthtnew");
+		dot_new_predicates_operands.add("A4_pzxthfnew");
+		dot_new_predicates_operands.add("A4_paslhtnew");
+		dot_new_predicates_operands.add("A4_paslhfnew");
+		dot_new_predicates_operands.add("A4_pasrhtnew");
+		dot_new_predicates_operands.add("A4_pasrhfnew");
+	}
+
 	static final Set<String> new_cmp_jumps;
 
 	static {
@@ -327,8 +591,8 @@ public class HexagonPcodeEmitPacked {
 		// any (non-NEWCMPJMP) instruction that reads a dot-new predicate should be
 		// handled after all other instructions, to give a "future" producer the chance
 		// to perform the store
-		if (HexagonPacketInfo.dot_new_predicates.containsKey(instr.getMnemonicString())
-				|| HexagonPacketInfo.dot_new_predicates_operands.contains(instr.getMnemonicString())) {
+		if (dot_new_predicates.containsKey(instr.getMnemonicString())
+				|| dot_new_predicates_operands.contains(instr.getMnemonicString())) {
 			return true;
 		}
 		return false;
@@ -366,29 +630,27 @@ public class HexagonPcodeEmitPacked {
 		newCmpJumpInstructions = new ArrayList<>();
 		instructionsContainingDotNewPredicates = new ArrayList<>();
 		while (insnIter.hasNext()) {
+			boolean part1 = false;
 			Instruction instr = insnIter.next();
-			if (isNewCmpJumpInstruction(instr)) {
-				newCmpJumpInstructions.add(instr);
-				continue;
-			}
 			if (doesInstructionContainDotNewPredicate(instr)) {
 				instructionsContainingDotNewPredicates.add(instr);
-				continue;
+				if (isNewCmpJumpInstruction(instr)) {
+					// resolve pcode for "part1" of new-comp jump instruction (the compare portion)
+					part1 = true;
+				} else {
+					// don't resolve this until a later phase
+					continue;
+				}
 			}
 
-			final_pcode.addAll(fixupPcode(instr));
-		}
-	}
-
-	void processNewCmpJumpInstructions() throws UnknownInstructionException {
-		for (Instruction instr : newCmpJumpInstructions) {
-			final_pcode.addAll(fixupPcode(instr));
+			final_pcode.addAll(fixupPcode(instr, part1, false));
 		}
 	}
 
 	void processInstructionsContainingDotNewPredicates() throws UnknownInstructionException {
 		for (Instruction instr : instructionsContainingDotNewPredicates) {
-			final_pcode.addAll(fixupPcode(instr));
+			// resolve pcode for "part2" of new-comp jump instruction (the new-jump portion)
+			final_pcode.addAll(fixupPcode(instr, false, true));
 		}
 	}
 
@@ -485,9 +747,48 @@ public class HexagonPcodeEmitPacked {
 
 	List<HexagonExternalBranch> branches;
 
-	List<PcodeOp> fixupPcode(Instruction instr) throws UnknownInstructionException {
-		List<PcodeOp> ops = Arrays.asList(instr.getPrototype().getPcode(instr.getInstructionContext(), null, null));
+	Set<Varnode> autoAndPredicatesWritten;
+
+	List<PcodeOp> fixupPcode(Instruction instr, boolean part1, boolean part2) throws UnknownInstructionException {
+		boolean insertedInstrumentationForAutoAndPredicate = false;
+		boolean hasRelativeCof = false;
+		boolean skipNextInstruction = false;
+
+		InstructionPrototype proto;
+		InstructionContext ctx;
+
+		if (isNewCmpJumpInstruction(instr) && (part1 || part2)) {
+			//
+			// p0=cmp.eq(Rs16,#U5); if (p0.new) jump:nt #r9:2
+			//
+			// part1: fWRITE_P0(f8BITSOF((RsV==UiV)))
+			//
+			// part2: if (fLSBNEW0) { fIMMEXT(riV); fBRANCH(fREAD_PC()+riV,COF_TYPE_JUMP); }
+			//
+			try {
+				ProcessorContext impl = new ProcessorContextImpl(program.getLanguage());
+				impl.setValue(part1Register, BigInteger.valueOf(part1 ? 1 : 0));
+				impl.setValue(part2Register, BigInteger.valueOf(part2 ? 1 : 0));
+
+				proto = program.getLanguage().parse(instr.getInstructionContext().getMemBuffer(), impl, false);
+				ctx = new PseudoInstruction(program, instr.getAddress(), proto,
+						instr.getInstructionContext().getMemBuffer(), impl);
+			} catch (ContextChangeException | InsufficientBytesException | AddressOverflowException e) {
+				String msg = "Unexpected exception when trying to break out new-cmp jump into parts" + e;
+				throw new UnknownInstructionException(msg);
+			}
+		} else {
+			proto = instr.getPrototype();
+			ctx = instr.getInstructionContext();
+		}
+
+		LinkedList<PcodeOp> ops = new LinkedList<>(Arrays.asList(proto.getPcode(ctx, null, null)));
+
 		for (int i = 0; i < ops.size(); i++) {
+			if (skipNextInstruction) {
+				skipNextInstruction = false;
+				continue;
+			}
 			PcodeOp op = ops.get(i);
 			if (isCallotherNewreg(op)) {
 				// new-value operand/dot-new predicate must have been written earlier in packet
@@ -504,8 +805,30 @@ public class HexagonPcodeEmitPacked {
 				if (op.getOutput() != null && op.getOutput().isRegister()) {
 					regsWrittenSoFar.add(op.getOutput());
 					if (isPredicateRegister(op.getOutput())) {
-						// TODO: handle auto-and predicate
-						op.setOutput(getScratchReg(instr, op.getOutput()));
+						// Section 6.1.3 in "Hexagon V66 Programmer’s Reference Manual"
+						// > If multiple compare instructions in a packet write to the same
+						// > predicate register, the result is the logical AND of the
+						// > individual compare results
+						Varnode prd = op.getOutput();
+						if (autoAndPredicatesWritten.contains(prd)) {
+							if (op.getOpcode() != PcodeOp.COPY) {
+								Varnode tmp = new Varnode(uniqueFactory.getNextUniqueAddress(), 1);
+								op.setOutput(tmp);
+								// next iter of the loop will fix this instruction up
+								PcodeOp auto_and = And(getScratchReg(instr, prd), tmp);
+								ops.add(i + 1, auto_and);
+								insertedInstrumentationForAutoAndPredicate = true;
+								skipNextInstruction = true;
+							} else {
+								// we can transform copies into auto-and predicates fairly easily
+								ops.set(i, And(getScratchReg(instr, prd), op.getInput(0)));
+							}
+						} else {
+							// first set of this predicate, perform the store "normally"
+							op.setOutput(getScratchReg(instr, op.getOutput()));
+							autoAndPredicatesWritten.add(prd);
+						}
+
 					} else {
 						op.setOutput(getScratchReg(instr, op.getOutput()));
 					}
@@ -526,6 +849,8 @@ public class HexagonPcodeEmitPacked {
 					PcodeOp hitOp = Copy(br.condVn, hit);
 					ops.set(i, hitOp);
 					branchNoInInsn++;
+				} else {
+					hasRelativeCof = true;
 				}
 				break;
 			case PcodeOp.CBRANCH:
@@ -535,9 +860,25 @@ public class HexagonPcodeEmitPacked {
 					PcodeOp hitOp = Copy(br.condVn, op.getInput(1));
 					ops.set(i, hitOp);
 					branchNoInInsn++;
+				} else {
+					hasRelativeCof = true;
 				}
 				break;
 			}
+		}
+
+		if (hasRelativeCof && insertedInstrumentationForAutoAndPredicate) {
+			// I don't think this is reachable since only newcmpjumps should write a
+			// predicate and then branch. Hence the runtime assert instead of throwing
+			// UnknownInstructionException
+			assert false;
+			String msg = "NYI: auto-and predicates required, but internal cof could not be fixed up " + instr;
+			if (part1) {
+				msg += " (1st part)";
+			} else if (part2) {
+				msg += " (2nd part)";
+			}
+			throw new AssertionError(msg);
 		}
 		return ops;
 	}
@@ -586,6 +927,8 @@ public class HexagonPcodeEmitPacked {
 
 	List<PcodeOp> final_pcode;
 
+	UniqueAddressFactory uniqueFactory;
+
 	public List<PcodeOp> getPcode(InstructionContext context, UniqueAddressFactory uniqueFactory)
 			throws UnknownInstructionException {
 		BigInteger pkt_start = program.getProgramContext()
@@ -598,10 +941,6 @@ public class HexagonPcodeEmitPacked {
 			throw new UnknownInstructionException("Attempting to get pcode from the middle of a packet");
 		}
 
-		minAddr = program.getAddressFactory().getDefaultAddressSpace().getAddress(pkt_start.longValue());
-
-		defaultSeqno = new SequenceNumber(minAddr, 0);
-
 		BigInteger pkt_next = program.getProgramContext().getValue(program.getProgramContext().getRegister("pkt_next"),
 				context.getAddress(), false);
 		if (pkt_next == null) {
@@ -609,17 +948,25 @@ public class HexagonPcodeEmitPacked {
 			throw new UnknownInstructionException("Packet not yet analyzed");
 		}
 
+		minAddr = program.getAddressFactory().getDefaultAddressSpace().getAddress(pkt_start.longValue());
+		defaultSeqno = new SequenceNumber(minAddr, 0);
+
 		pktNext = program.getAddressFactory().getDefaultAddressSpace().getAddress(pkt_next.longValue());
+
 		maxAddr = pktNext.subtract(1);
 
 		packetSize = pkt_next.subtract(pkt_start).intValue();
 
 		AddressSet addrSet = new AddressSet(minAddr, maxAddr);
 
+		this.uniqueFactory = uniqueFactory;
+
 		regTempSpace = new HexagonRegisterScratchSpace(program, uniqueFactory);
 		regTempSpaceWrite = new HexagonRegisterScratchSpace(program, uniqueFactory);
 
 		regsWrittenSoFar = new HashSet<>(); // debugging tool
+
+		autoAndPredicatesWritten = new HashSet<>();
 
 		final_pcode = new ArrayList<PcodeOp>();
 
@@ -629,55 +976,91 @@ public class HexagonPcodeEmitPacked {
 		// 1. First, all registers used into the packet are spilled into temporaries in
 		// either regTempSpace or regTempSpaceWrite
 		//
+
+		initializeTemporaryRegisters(program.getListing().getInstructions(addrSet, true));
+
+		//
 		// 2. All branches in the packet are recorded in the order they appear in the
 		// packet. Record and initialize a "branch var" that records whether that branch
 		// is taken
 		//
-		// 3. Copy pcode for all instructions (besides instructions containing dot-new
-		// predicates). Replace all registers with their appropriate temporaries
+		// Order of jumps in the packet is important, because this is the order
+		// jumps are resolved in (e.g. if one branch is conditional, and the
+		// other branch is unconditional, the jump in the earlier instruction is
+		// taken
 		//
-		// 4. Copy pcode for all newcmpjump instructions
+
+		recordBranchesAndValidateImplementedPcode(program.getListing().getInstructions(addrSet, true), uniqueFactory);
+
 		//
-		// 5. Copy pcode for all other instructions containing dot-new predicates
+		// 3. Copy pcode for all instructions that don't contain dot-new predicates.
+		// Replace all registers with their appropriate temporaries
+		//
+		// N.B. there is a class of instructions, newcmpjump instructions of the form
+		//
+		// P0 = cmp.eq(R18,#0x0); if (P0.new) jump:nt 1f
+		//
+		// Such instructions both read and write a predicate register. They must be
+		// split into part1 (the compare part) and part2 (the dot-new jump part). The
+		// compare part's pcode is copied in this step. The dot-new jump part is handled
+		// below.
+		//
+		// This is required for two reasons. Instructions containing dot-new
+		// predicates can be earlier in the packet than the corresponding store:
+		//
+		// { if (!P0.new) r0 = #41
+		// P0 = cmp.eq(R18,#0x0); if (P0.new) jump:nt 1f }
+		//
+		// We need to handle all dot-new predicates after the rest of the instructions
+		// in the packet
+		//
+		// auto-and predicates presents an additional challenge. This is also valid:
+		//
+		// { p0 = cmp.eq(r5, #7); if (p0.new) jump:nt 1f
+		// p0 = cmp.eq(r6, #7) }
+		//
+		// p0 is written twice in the same packet, so it's value at the end of the
+		// packet is boolean AND'd: (r5 == #7) && (r6 == #7). p0.new also receives this
+		// value, despite occuring in the packet before the second store to p0.
+		//
+		// This requires us to split up newcmpjump instructions into the cmp and dot-new
+		// parts, instead of simply processing them after all other instructions
+		//
+		// N.B. new-value operands do not have the problem stated above, because
+		// new-value operands are encoded with a constant indicating a *previous*
+		// instruction in the packet. So the new-value producer must come earlier in the
+		// packet
+		//
+		// N.B. Auto-and predicates are supported
+		//
+		// N.B. Code assumes that there is at most one external branch in every
+		// instruction, not including fallthrough. endloop01 is the only exception with
+		// two external jumps. FIXME there is a subtle bug causing improper
+		// decompilation somewhere with endloop01, see testHwLoop01 in
+		// HexagonPacketTestDecompilation
+		//
+		// N.B. Emitted pcode respects flow overrides and handles them appropriately
+		//
+
+		processInstructionsBesidesDotNewPredicates(program.getListing().getInstructions(addrSet, true));
+
+		//
+		// 5. Copy pcode for all other instructions containing dot-new predicates,
+		// including the dot-new jump parts (part2) of newcmpjump instructions
+		//
+
+		processInstructionsContainingDotNewPredicates();
+
 		//
 		// 6. Resolve all register stores
+		//
+
+		resolveAllStores();
+
 		//
 		// 7. Resolve the correct branch destination in priority of the order each
 		// branch appears in the packet
 		//
-		// Some interesting comments:
-		//
-		// - Steps 4 and 5 are required because dot-new predicates (p0.new) can occur
-		// anywhere in the packet, before the corresponding store to the predicate. So
-		// we process all instructions that contain a dot-new predicate after the rest
-		// of the instructions have been processed.
-		//
-		// Note that NewCmpJump instructions are those which both write a predicate
-		// register and read that dot-new predicate
-		//
-		// - new-value operands do not have the problem above, because new-value
-		// operands are encoded with a constant indicating a *previous* instruction in
-		// the packet. So the new-value producer must come earlier in the packet
-		//
-		// - TODO: Currently auto-and predicates are not supported
-		//
-		// - Code assumes that there is at most one external branch in every
-		// instruction, not including fallthrough. endloop01 is the only exception
-		//
-		// - Emitted pcode respects flow overrides and handles them appropriately
-		//
-
-		initializeTemporaryRegisters(program.getListing().getInstructions(addrSet, true));
-
-		recordBranchesAndValidateImplementedPcode(program.getListing().getInstructions(addrSet, true), uniqueFactory);
-
-		processInstructionsBesidesDotNewPredicates(program.getListing().getInstructions(addrSet, true));
-
-		processNewCmpJumpInstructions();
-
-		processInstructionsContainingDotNewPredicates();
-
-		resolveAllStores();
 
 		resolveAllBranches();
 
@@ -713,7 +1096,7 @@ public class HexagonPcodeEmitPacked {
 	}
 
 	PcodeOp And(Varnode dst, Varnode src) {
-		return new PcodeOp(defaultSeqno, PcodeOp.COPY, new Varnode[] { dst, src }, dst);
+		return new PcodeOp(defaultSeqno, PcodeOp.BOOL_AND, new Varnode[] { dst, src }, dst);
 	}
 
 	Varnode Constant(int val) {
